Rename "transform" subcommand to "run"
